refactor: split api bindings module
