+++ pkgs/standards/autoapi/autoapi/v3/bindings/api/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from .common import AttrDict, _default_prefix, _mount_router  # noqa: F401
+from .include import include_model, include_models, _seed_security_and_deps  # noqa: F401
+from .rpc import rpc_call
+
+__all__ = ["include_model", "include_models", "rpc_call"]
