+++ pkgs/standards/autoapi/autoapi/v3/bindings/api/include.py
@@ -1,0 +1,200 @@
+from __future__ import annotations
+
+import logging
+from types import SimpleNamespace
+from typing import Any, Dict, Sequence, Tuple
+
+from .common import (
+    ApiLike,
+    _default_prefix,
+    _ensure_api_ns,
+    _has_include_router,
+    _mount_router,
+    _resource_name,
+)
+from .resource_proxy import _ResourceProxy
+from .. import model as _binder
+from ...config.constants import (
+    AUTOAPI_AUTH_DEP_ATTR,
+    AUTOAPI_AUTHORIZE_ATTR,
+    AUTOAPI_GET_DB_ATTR,
+    AUTOAPI_REST_DEPENDENCIES_ATTR,
+    AUTOAPI_RPC_DEPENDENCIES_ATTR,
+    AUTOAPI_ALLOW_ANON_ATTR,
+)
+from ...engine import resolver as _resolver
+
+logger = logging.getLogger(__name__)
+
+
+# --- keep as helper, no behavior change to transports/kernel ---
+def _seed_security_and_deps(api: Any, model: type) -> None:
+    """
+    Copy API-level dependency hooks onto the model so downstream binders can use them.
+    - __autoapi_get_db__             : DB dep (FastAPI Depends-compatible)
+    - __autoapi_auth_dep__           : auth dependency (returns user or raises 401)
+    - __autoapi_authorize__          : callable(request, model, alias, payload, user)→None/raise 403
+    - __autoapi_rest_dependencies__  : list of extra dependencies for REST (e.g., rate-limits)
+    - __autoapi_rpc_dependencies__   : list of extra dependencies for JSON-RPC router
+    """
+    # DB deps
+    prov = _resolver.resolve_provider(api=api)
+    if prov is not None:
+        setattr(model, AUTOAPI_GET_DB_ATTR, prov.get_db)
+
+    # Authn (prefer optional dep when available)
+    auth_dep = None
+    if getattr(api, "_optional_authn_dep", None):
+        auth_dep = api._optional_authn_dep
+    elif getattr(api, "_allow_anon", True) is False and getattr(api, "_authn", None):
+        auth_dep = api._authn
+    elif getattr(api, "_authn", None):
+        auth_dep = api._authn
+    if auth_dep is not None:
+        setattr(model, AUTOAPI_AUTH_DEP_ATTR, auth_dep)
+
+    # Allow anonymous verbs
+    allow_attr = getattr(model, AUTOAPI_ALLOW_ANON_ATTR, None)
+    if allow_attr:
+        verbs = allow_attr() if callable(allow_attr) else allow_attr
+        for v in verbs:
+            api._allow_anon_ops.add(f"{model.__name__}.{v}")
+
+    # Authz
+    if getattr(api, "_authorize", None):
+        setattr(model, AUTOAPI_AUTHORIZE_ATTR, api._authorize)
+
+    # Extra deps (router-level only; never part of runtime plan)
+    if getattr(api, "rest_dependencies", None):
+        setattr(model, AUTOAPI_REST_DEPENDENCIES_ATTR, list(api.rest_dependencies))
+    if getattr(api, "rpc_dependencies", None):
+        setattr(model, AUTOAPI_RPC_DEPENDENCIES_ATTR, list(api.rpc_dependencies))
+
+
+def _attach_to_api(api: ApiLike, model: type) -> None:
+    """
+    Attach the model’s bound namespaces to the api facade.
+    """
+    _ensure_api_ns(api)
+
+    mname = model.__name__
+    rname = _resource_name(model)
+    rtitle = rname[:1].upper() + rname[1:]
+
+    # Index model object
+    api.models[mname] = model
+    api.tables[mname] = getattr(model, "__table__", None)
+
+    # Direct references to model namespaces
+    setattr(api.schemas, mname, getattr(model, "schemas", SimpleNamespace()))
+    setattr(api.handlers, mname, getattr(model, "handlers", SimpleNamespace()))
+    setattr(api.hooks, mname, getattr(model, "hooks", SimpleNamespace()))
+    rpc_ns = getattr(model, "rpc", SimpleNamespace())
+    setattr(api.rpc, mname, rpc_ns)
+    if rtitle != mname:
+        setattr(api.rpc, rtitle, rpc_ns)
+    # rest (router lives on model.rest.router)
+    rest_ns = getattr(api, "rest")
+    setattr(
+        rest_ns,
+        mname,
+        SimpleNamespace(
+            router=getattr(getattr(model, "rest", SimpleNamespace()), "router", None)
+        ),
+    )
+    # also keep a flat routers dict for quick access
+    api.routers[mname] = getattr(
+        getattr(model, "rest", SimpleNamespace()), "router", None
+    )
+
+    # Table metadata (introspection only)
+    api.columns[mname] = tuple(getattr(model, "columns", ()))
+    api.table_config[mname] = dict(getattr(model, "table_config", {}) or {})
+
+    # Core helper proxies (now aware of API for DB resolution precedence)
+    core_proxy = _ResourceProxy(model, api=api)
+    setattr(api.core, mname, core_proxy)
+    if rtitle != mname:
+        setattr(api.core, rtitle, core_proxy)
+
+    core_raw_proxy = _ResourceProxy(model, serialize=False, api=api)
+    setattr(api.core_raw, mname, core_raw_proxy)
+    if rtitle != mname:
+        setattr(api.core_raw, rtitle, core_raw_proxy)
+
+
+def include_model(
+    api: ApiLike,
+    model: type,
+    *,
+    app: Any | None = None,
+    prefix: str | None = None,
+    mount_router: bool = True,
+) -> Tuple[type, Any]:
+    """
+    Bind a model (if not already bound), mount its REST router, and attach all namespaces to `api`.
+
+    Args:
+        api: An arbitrary facade object; we’ll attach containers onto it if missing.
+        model: The SQLAlchemy model (table class).
+        app: Optional FastAPI app or Router (anything with `include_router`).
+             Routers are always mounted on `api.router`; if provided, we also
+             mount onto this `app` (or `api.app` when not given).
+        prefix: Optional mount prefix. When None, defaults to `/{ModelClassName}` or
+                `/{__resource__}` if set on the model.
+        mount_router: If False, we skip mounting onto the host app but still bind
+            the router under `api.router`/`api.rest`/`api.routers`.
+
+    Returns:
+        (model, router) – the model class and its Router (or None if not present).
+    """
+    # 0) seed deps/security so binders can see them (transport-level only)
+    _seed_security_and_deps(api, model)
+
+    # 1) Build/bind model namespaces (idempotent)
+    _binder.bind(model)
+
+    # 2) Pick a router & mount prefix
+    router = getattr(getattr(model, "rest", SimpleNamespace()), "router", None)
+    if prefix is None:
+        prefix = _default_prefix(model)
+
+    # 3) Always bind model router to the API object when possible
+    root_router = api if _has_include_router(api) else getattr(api, "router", None)
+    if router is not None:
+        _mount_router(root_router, router, prefix=prefix)
+
+    # Optionally mount onto a host app
+    target_app = app or getattr(api, "app", None)
+    if mount_router and router is not None:
+        _mount_router(target_app, router, prefix=prefix)
+
+    # 4) Attach all namespaces onto api
+    _attach_to_api(api, model)
+
+    logger.debug("bindings.api: included %s at prefix %s", model.__name__, prefix)
+    return model, router
+
+
+def include_models(
+    api: ApiLike,
+    models: Sequence[type],
+    *,
+    app: Any | None = None,
+    base_prefix: str | None = None,
+    mount_router: bool = True,
+) -> Dict[str, Any]:
+    """
+    Convenience helper to include multiple models.
+
+    If ``base_prefix`` is provided, each model's router is mounted under that
+    prefix. The model router itself already has its own `/{resource}` prefix.
+    """
+    results: Dict[str, Any] = {}
+    for mdl in models:
+        px = base_prefix.rstrip("/") if base_prefix else None
+        _, router = include_model(
+            api, mdl, app=app, prefix=px, mount_router=mount_router
+        )
+        results[mdl.__name__] = router
+    return results
